requirements.txt
python-dotenv
azure-identity
azure-storage-file-datalake
azure-cli
pandas
tqdm
pip-system-certs
coverage
pytest